Move duplicate code into LabelTrait
