+++ models/DataObject/ClassDefinition/Layout/Fieldset.php
@@ -17,71 +17,16 @@
 namespace Pimcore\Model\DataObject\ClassDefinition\Layout;
 
 use Pimcore\Model;
+use Pimcore\Model\DataObject\ClassDefinition\Layout\Traits\LabelTrait;
 
 class Fieldset extends Model\DataObject\ClassDefinition\Layout
 {
+    use LabelTrait;
+
     /**
      * Static type of this element
      *
      * @var string
      */
     public $fieldtype = 'fieldset';
-
-    /**
-     * Width of input field labels
-     *
-     * @var int
-     */
-    public $labelWidth = 100;
-
-    /**
-     * Align of input field labels
-     *
-     * @var string
-     */
-    public $labelAlign = 'left';
-
-    /**
-     * @param int $labelWidth
-     *
-     * @return $this
-     */
-    public function setLabelWidth($labelWidth)
-    {
-        if (!empty($labelWidth)) {
-            $this->labelWidth = intval($labelWidth);
-        }
-
-        return $this;
-    }
-
-    /**
-     * @return int
-     */
-    public function getLabelWidth()
-    {
-        return $this->labelWidth;
-    }
-
-    /**
-     * @param string $labelAlign
-     *
-     * @return $this
-     */
-    public function setLabelAlign($labelAlign): self
-    {
-        if (!empty($labelAlign)) {
-            $this->labelAlign = $labelAlign;
-        }
-
-        return $this;
-    }
-
-    /**
-     * @return string
-     */
-    public function getLabelAlign()
-    {
-        return $this->labelAlign;
-    }
 }
